+++ src/models.rs
@@ -19,10 +19,10 @@ pub use transaction::{
 
 use crate::parser::value_expr::build_root_node_from_expression;
 use crate::parser::{tokenizers, value_expr};
+use crate::List;
 use crate::{error::EmptyLedgerFileError, parser::ParsedLedger};
 use crate::{filter::filter_expression, CommonOpts};
 use crate::{models::transaction::Cost, parser::Tokenizer};
-use crate::{GenericError, List};
 use num::BigInt;
 use std::cell::RefCell;
 use std::rc::Rc;
@@ -373,7 +373,7 @@ impl ParsedLedger {
     fn _transaction_to_ledger(
         &self,
         parsed: &Transaction<tokenizers::transaction::RawPosting>,
-    ) -> Result<TransactionTransformer, GenericError> {
+    ) -> Result<TransactionTransformer, Box<dyn std::error::Error>> {
         let mut automated_transactions = vec![];
         let mut prices = vec![];
         let mut transactions = vec![];